refactor: move cypress commands to commands.js
