+++ cypress/integration/racingcar.spec.js
@@ -1,47 +1,5 @@
-import { SELECTOR } from '../../src/constants/selector.js';
-
 describe('구현 결과가 요구사항과 일치해야 한다.', () => {
   const baseUrl = "../../index.html";
-
-  before(() => {
-    Cypress.Commands.add('normalWorking', (carNames, racingCount) => {
-      cy.get(SELECTOR.CAR_NAME_INPUT).type(carNames);
-      cy.get(SELECTOR.CAR_NAME_BUTTON).click();
-  
-      cy.get(SELECTOR.RACE_COUNT_INPUT).type(racingCount);
-      cy.get(SELECTOR.RACE_COUNT_BUTTON).click();
-    });
-
-    Cypress.Commands.add('showCarNameAlert', (invalidInput) => {
-      const alertStub = cy.stub();
-      cy.on("window:alert", alertStub);
-    
-      cy.get(SELECTOR.CAR_NAME_INPUT).type(invalidInput);
-    
-      cy.get(SELECTOR.CAR_NAME_BUTTON)
-        .click()
-        .then(() => {
-          expect(alertStub).to.be.called;
-      });
-    });
-
-    Cypress.Commands.add('showRaceCountAlert', (carNameInput, invalidInput) => {
-      const alertStub = cy.stub();
-      cy.get(SELECTOR.CAR_NAME_INPUT).type(carNameInput);
-      cy.get(SELECTOR.CAR_NAME_BUTTON).click();
-
-      cy.on("window:alert", alertStub);
-
-      cy.get(SELECTOR.RACE_COUNT_INPUT).type(invalidInput);
-
-      cy.get(SELECTOR.RACE_COUNT_BUTTON)
-        .click()
-        .then(() => {
-          expect(alertStub).to.be.called;
-      });
-    });
-
-  });
 
   beforeEach(() => {
     cy.visit(baseUrl);
