+++ cypress/support/commands.js
@@ -7,19 +7,42 @@
 // commands please read more here:
 // https://on.cypress.io/custom-commands
 // ***********************************************
-//
-//
-// -- This is a parent command --
-// Cypress.Commands.add('login', (email, password) => { ... })
-//
-//
-// -- This is a child command --
-// Cypress.Commands.add('drag', { prevSubject: 'element'}, (subject, options) => { ... })
-//
-//
-// -- This is a dual command --
-// Cypress.Commands.add('dismiss', { prevSubject: 'optional'}, (subject, options) => { ... })
-//
-//
-// -- This will overwrite an existing command --
-// Cypress.Commands.overwrite('visit', (originalFn, url, options) => { ... })
+
+import { SELECTOR } from '../../src/constants/selector.js';
+
+Cypress.Commands.add('normalWorking', (carNames, racingCount) => {
+    cy.get(SELECTOR.CAR_NAME_INPUT).type(carNames);
+    cy.get(SELECTOR.CAR_NAME_BUTTON).click();
+
+    cy.get(SELECTOR.RACE_COUNT_INPUT).type(racingCount);
+    cy.get(SELECTOR.RACE_COUNT_BUTTON).click();
+});
+
+Cypress.Commands.add('showCarNameAlert', (invalidInput) => {
+    const alertStub = cy.stub();
+    cy.on("window:alert", alertStub);
+
+    cy.get(SELECTOR.CAR_NAME_INPUT).type(invalidInput);
+
+    cy.get(SELECTOR.CAR_NAME_BUTTON)
+        .click()
+        .then(() => {
+        expect(alertStub).to.be.called;
+    });
+});
+
+Cypress.Commands.add('showRaceCountAlert', (carNameInput, invalidInput) => {
+    const alertStub = cy.stub();
+    cy.get(SELECTOR.CAR_NAME_INPUT).type(carNameInput);
+    cy.get(SELECTOR.CAR_NAME_BUTTON).click();
+
+    cy.on("window:alert", alertStub);
+
+    cy.get(SELECTOR.RACE_COUNT_INPUT).type(invalidInput);
+
+    cy.get(SELECTOR.RACE_COUNT_BUTTON)
+        .click()
+        .then(() => {
+        expect(alertStub).to.be.called;
+    });
+});
